Streamline magic,strength => power
There was little reason for a magic class to go for strength, or a physical class to go for magic.
As a result, all classes had a dead stat. By tying directly into the secondary stat 'power',
we should be able to streamline damage per second etc logic as well.

Also included is an overall streamlining of enemy weapons. Balance massively TODO